Rename lambda to test-lambda-riffraff
