+++ packages/cdk/lib/test-lambda-riffraff.test.ts
@@ -1,21 +1,17 @@
 import { App } from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
-import { ExampleTypescriptLambda } from './example-typescript-lambda';
+import { TestTypescriptLambda } from './test-lambda-riffraff';
 
-describe('The ExampleTypescriptLambda stack', () => {
+describe('The TestTypescriptLambda stack', () => {
 	it('matches the snapshot', () => {
 		const app = new App();
-		const stack = new ExampleTypescriptLambda(
-			app,
-			'ExampleTypescriptLambda-1',
-			{
-				stack: 'playground',
-				stage: 'TEST',
-				env: {
-					region: 'eu-west-1',
-				},
+		const stack = new TestTypescriptLambda(app, 'TestTypescriptLambda', {
+			stack: 'playground',
+			stage: 'TEST',
+			env: {
+				region: 'eu-west-1',
 			},
-		);
+		});
 		const template = Template.fromStack(stack);
 
 		/**
